customer_dashboard: Extract form logic into useEditInfoForm hook
