+++ apps/customer_dashboard/src/components/edit_info_form/edit_info_form.tsx
@@ -1,14 +1,6 @@
 "use client";
 
-import {
-  type ChangeEvent,
-  useRef,
-  useState,
-  type DragEvent,
-  type FC,
-} from "react";
-import { useRouter } from "next/navigation";
-import { useQueryClient } from "@tanstack/react-query";
+import type { FC } from "react";
 import { Input } from "@oko-wallet/oko-common-ui/input";
 import { Button } from "@oko-wallet/oko-common-ui/button";
 import { PlusIcon } from "@oko-wallet/oko-common-ui/icons/plus";
@@ -16,228 +8,55 @@ import { XCloseIcon } from "@oko-wallet/oko-common-ui/icons/x_close";
 import type { CustomerTheme } from "@oko-wallet/oko-types/customers";
 import { Typography } from "@oko-wallet/oko-common-ui/typography";
 
-import { useCustomerInfo } from "@oko-wallet-ct-dashboard/hooks/use_customer_info";
-import { useAppState } from "@oko-wallet-ct-dashboard/state";
-import { requestUpdateCustomerInfo } from "@oko-wallet-ct-dashboard/fetch/customers";
+import { useEditInfoForm } from "./use_edit_info_form";
 import styles from "./edit_info_form.module.scss";
 
 const THEME_OPTIONS: CustomerTheme[] = ["light", "dark", "system"];
 
 export const EditInfoForm: FC = () => {
-  const router = useRouter();
-  const queryClient = useQueryClient();
-  const customer = useCustomerInfo();
-  const token = useAppState((state) => state.token);
-
-  const fileInputRef = useRef<HTMLInputElement>(null);
-
-  const [label, setLabel] = useState(customer.data?.label ?? "");
-  const [url, setUrl] = useState(customer.data?.url ?? "");
-  const [theme, setTheme] = useState<CustomerTheme>(
-    customer.data?.theme ?? "system",
-  );
-
-  const [logoFile, setLogoFile] = useState<File | null>(null);
-  const [previewUrl, setPreviewUrl] = useState<string | null>(
-    customer.data?.logo_url ?? null,
-  );
-  const [isLoading, setIsLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-  const [isDragging, setIsDragging] = useState(false);
-  const [shouldDeleteLogo, setShouldDeleteLogo] = useState(false);
-
-  // Validate image dimensions (128x128)
-  const validateImageDimensions = (file: File): Promise<boolean> => {
-    return new Promise((resolve) => {
-      const img = new Image();
-      img.onload = () => {
-        URL.revokeObjectURL(img.src);
-        resolve(img.width === 128 && img.height === 128);
-      };
-      img.onerror = () => {
-        URL.revokeObjectURL(img.src);
-        resolve(false);
-      };
-      img.src = URL.createObjectURL(file);
-    });
-  };
-
-  // File upload handler with proper validation
-  const handleLogoUpload = async (file: File) => {
-    // Check file type (no SVG, no GIF)
-    const allowedTypes = ["image/png", "image/jpeg", "image/jpg", "image/webp"];
-    if (!allowedTypes.includes(file.type)) {
-      setError(
-        "Only PNG, JPG, and WebP image files are allowed. SVG and GIF are not supported.",
-      );
-      return;
-    }
-
-    // File size validation (1MB limit)
-    if (file.size > 1 * 1024 * 1024) {
-      setError("File size must be under 1 MB.");
-      return;
-    }
-
-    // Validate dimensions (128x128)
-    const isValidDimensions = await validateImageDimensions(file);
-    if (!isValidDimensions) {
-      setError("Image must be exactly 128×128 pixels.");
-      return;
-    }
-
-    setError(null);
-    setShouldDeleteLogo(false);
-
-    const reader = new FileReader();
-    reader.onload = (e) => {
-      setPreviewUrl(e.target?.result as string);
-    };
-    reader.readAsDataURL(file);
-
-    setLogoFile(file);
-  };
-
-  // File remove handler - completely remove logo
-  const handleLogoRemove = () => {
-    setPreviewUrl(null);
-    setLogoFile(null);
-    setShouldDeleteLogo(true); // Mark for deletion
-    setError(null);
-    if (fileInputRef.current) {
-      fileInputRef.current.value = "";
-    }
-  };
-
-  const handleFileSelect = (e: ChangeEvent<HTMLInputElement>) => {
-    const file = e.target.files?.[0];
-    if (file) {
-      handleLogoUpload(file);
-    }
-  };
-
-  const handleUploadClick = () => {
-    fileInputRef.current?.click();
-  };
-
-  // Drag and drop handlers
-  const handleDragEnter = (e: DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragging(true);
-  };
-
-  const handleDragLeave = (e: DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragging(false);
-  };
-
-  const handleDragOver = (e: DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-  };
-
-  const handleDrop = (e: DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragging(false);
-
-    const file = e.dataTransfer?.files?.[0];
-    if (file) {
-      handleLogoUpload(file);
-    }
-  };
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-
-    if (!token) {
-      setError("Please log in to continue.");
-      return;
-    }
-
-    // Check for changes
-    const hasLabelChange = label !== customer.data?.label;
-    const hasUrlChange = url !== (customer.data?.url ?? "");
-    const hasLogoChange = logoFile !== null || shouldDeleteLogo;
-    const hasThemeChange = theme !== customer.data?.theme;
-
-    if (!hasLabelChange && !hasUrlChange && !hasLogoChange && !hasThemeChange) {
-      setError("No changes to save.");
-      return;
-    }
-
-    if (!url || url.trim() === "") {
-      setError("App URL is required.");
-      return;
-    }
-
-    if (hasUrlChange && !validateUrl(url)) {
-      setError("App URL format is invalid.");
-      return;
-    }
-
-    setIsLoading(true);
-    setError(null);
-
-    try {
-      const result = await requestUpdateCustomerInfo({
-        token,
-        label: hasLabelChange ? label : undefined,
-        url: hasUrlChange ? url : undefined,
-        logoFile: logoFile,
-        theme: hasThemeChange ? theme : undefined,
-        deleteLogo: shouldDeleteLogo,
-      });
-
-      if (result.success) {
-        // Invalidate query to refetch customer info
-        await queryClient.invalidateQueries({ queryKey: ["customer"] });
-
-        // Reset state
-        setLogoFile(null);
-
-        // Navigate back or show success message
-        alert("Information updated successfully!");
-        router.back();
-      } else {
-        setError(result.msg ?? "Failed to update information.");
-      }
-    } catch (err) {
-      setError("An error occurred while updating.");
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  const hasChanges =
-    label !== customer.data?.label ||
-    url !== (customer.data?.url ?? "") ||
-    logoFile !== null ||
-    shouldDeleteLogo ||
-    theme !== customer.data?.theme;
+  const {
+    onSubmit,
+    register,
+    errors,
+    isLoading,
+    hasChanges,
+    theme,
+    setTheme,
+    fileInputRef,
+    previewUrl,
+    isDragging,
+    handleFileSelect,
+    handleUploadClick,
+    handleLogoRemove,
+    handleDragEnter,
+    handleDragLeave,
+    handleDragOver,
+    handleDrop,
+    resetErrors,
+  } = useEditInfoForm();
 
   return (
-    <form onSubmit={handleSubmit} className={styles.form}>
+    <form onSubmit={onSubmit} className={styles.form}>
       {/* Team Name Input */}
       <Input
         label="Team Name"
-        value={label}
-        onChange={(e) => setLabel(e.target.value)}
+        {...register("label")}
         disabled={isLoading}
         placeholder="Team Name"
         className={styles.input}
+        error={errors.label?.message}
+        onFocus={() => resetErrors("label")}
       />
 
       {/* App URL Input */}
       <Input
         label="App URL"
-        value={url}
-        onChange={(e) => setUrl(e.target.value)}
+        {...register("url")}
         disabled={isLoading}
         placeholder="https://example.com"
         className={styles.input}
+        error={errors.url?.message}
+        onFocus={() => resetErrors("url")}
       />
 
       <div className={styles.themeSection}>
@@ -343,7 +162,7 @@ export const EditInfoForm: FC = () => {
       </div>
 
       {/* Error message */}
-      {error && <div className={styles.error}>{error}</div>}
+      {errors.root && <div className={styles.error}>{errors.root.message}</div>}
 
       {/* Submit button */}
       <Button
@@ -356,12 +175,3 @@ export const EditInfoForm: FC = () => {
     </form>
   );
 };
-
-function validateUrl(url: string): boolean {
-  try {
-    const urlObj = new URL(url);
-    return urlObj.protocol === "https:" || urlObj.protocol === "http:";
-  } catch {
-    return false;
-  }
-}
